refactor: fix account details retrieval in Account class
The code changes fix the issue with retrieving account details in the Account class. Previously, there was an incorrect concatenation of the 'money_value' and 'money_type' fields. This has been corrected by separating the fields and assigning the correct values. The updated code ensures accurate retrieval of account details.